Remove challenges window, localization updates
- Challenges are now integrated into the main window
- Localizations for challenges are now done.
