+++ SitStandTimer/ChallengeDescriptions.xcstrings
@@ -9,6 +9,18 @@
             "state" : "translated",
             "value" : "Stand on one foot for 15 seconds."
           }
+        },
+        "es" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Párate sobre un pie durante 15 segundos."
+          }
+        },
+        "fr" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Tenez-vous debout sur un pied pendant 15 secondes."
+          }
         }
       }
     },
@@ -19,6 +31,18 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "Do 15 jumping jacks."
+          }
+        },
+        "es" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Haz 15 saltos de tijera."
+          }
+        },
+        "fr" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Faites 15 sauts avec écart."
           }
         }
       }
@@ -31,6 +55,18 @@
             "state" : "translated",
             "value" : "Do 10 pushups against your desk."
           }
+        },
+        "es" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Haz 10 flexiones contra tu escritorio."
+          }
+        },
+        "fr" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Faites 10 pompes contre votre bureau."
+          }
         }
       }
     },
@@ -42,6 +78,18 @@
             "state" : "translated",
             "value" : "Perform 10 squats."
           }
+        },
+        "es" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Realiza 10 sentadillas."
+          }
+        },
+        "fr" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Effectuez 10 squats."
+          }
         }
       }
     },
@@ -52,6 +100,18 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "Do a 30-second full-body stretch."
+          }
+        },
+        "es" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Haz un estiramiento de cuerpo completo de 30 segundos."
+          }
+        },
+        "fr" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Faites un étirement de tout le corps pendant 30 secondes."
           }
         }
       }
